refactor: Migrate orbit-dashboard tests to per-module tests/ layout [ORB-00224]
