+++ crates/orbit-dashboard/src/api/tests/frictions.rs
@@ -10,7 +10,7 @@ use serde_json::{Value, json};
 use tower::ServiceExt;
 
 use super::super::router;
-use super::super::test_support::body_json;
+use super::test_support::body_json;
 
 fn seed_friction(runtime: &OrbitRuntime, body: &str, tags: &[&str]) -> Value {
     runtime
