+++ crates/orbit-dashboard/src/api/tests/learnings.rs
@@ -13,7 +13,7 @@ use serde_json::{Value, json};
 use tower::ServiceExt;
 
 use super::super::router;
-use super::super::test_support::body_json;
+use super::test_support::body_json;
 
 fn seed_learning(runtime: &OrbitRuntime, summary: &str) -> Learning {
     runtime
