+++ crates/orbit-dashboard/src/api/tests/tasks.rs
@@ -11,8 +11,9 @@ use orbit_core::{OrbitRuntime, TaskComplexity, TaskStatus};
 use serde_json::{Value, json};
 use tower::ServiceExt;
 
+use super::super::router;
+use super::super::tasks::*;
 use super::test_support::body_json;
-use super::*;
 
 fn seed_task_with_artifact(runtime: &OrbitRuntime) -> orbit_core::Task {
     let task = runtime
