+++ crates/orbit-dashboard/src/log_format/tests/mod.rs
@@ -1,0 +1,1 @@
+mod log_format;
